backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
PyJWT==2.8.0
django-cors-headers==4.3.1